Fix compiler warnings
-Wimplicit-int-float-conversion (Export::createSpinbutton)
-Wmisleading-indentation
-Wrange-loop-construct (init_extended)
-Wreturn-type (SPIFontSize::relative_fraction)
-Wunused-function
-Wunused-label (lpe_shape_revert_stroke_and_fill)
-Wunused-lambda-capture (text_categorize_refs)
-Wunused-private-field
-Wunused-variable

+++ src/ui/dialog/export.cpp
@@ -514,8 +514,8 @@ void Export::set_default_filename () {
     }
 }
 
-Glib::RefPtr<Gtk::Adjustment> Export::createSpinbutton( gchar const * /*key*/, float val, float min, float max,
-        float step, float page,
+Glib::RefPtr<Gtk::Adjustment> Export::createSpinbutton( gchar const * /*key*/,
+        double val, double min, double max, double step, double page,
         Gtk::Grid *t, int x, int y,
         const Glib::ustring& ll, const Glib::ustring& lr,
         int digits, unsigned int sensitive,
